pyproject.toml
[project]
name = "notionclient"
version = "0.1.0"
description = "Connect to the Notion API reference and create an another API to server a personalized response"
authors = [
    {name = "pyrec-dev"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "notion-client (>=2.4.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
